+++ src/lib.rs
@@ -9,6 +9,7 @@ mod blocks;
 mod changes;
 mod del;
 mod error;
+mod hb;
 mod keys;
 pub mod prefixed;
 mod put;
@@ -19,24 +20,24 @@ mod tree;
 use std::{
     fmt::Debug,
     ops::{Range, RangeBounds},
-    path::Path,
     sync::Arc,
 };
 
-use derive_builder::Builder;
-use hypercore::AppendOutcome;
-use prefixed::{Prefixed, PrefixedConfig};
 use prost::{bytes::Buf, DecodeError, Message};
 use random_access_storage::RandomAccess;
 use tokio::sync::RwLock;
 use tracing::trace;
 
 use blocks::Blocks;
-use error::HyperbeeError;
-use messages::{header::Metadata, yolo_index, YoloIndex};
-use traverse::{Traverse, TraverseConfig};
+use messages::{yolo_index, YoloIndex};
+
 use tree::Tree;
 
+pub use error::HyperbeeError;
+pub use hb::{Hyperbee, HyperbeeBuilder, HyperbeeBuilderError};
+pub use messages::header::Metadata;
+
+// TODO document
 pub trait CoreMem: RandomAccess + Debug + Send {}
 impl<T: RandomAccess + Debug + Send> CoreMem for T {}
 
@@ -49,26 +50,29 @@ fn min_keys(max_keys: usize) -> usize {
     max_keys >> 1
 }
 
-// TODO make not pub
 #[derive(Clone, Debug)]
 /// Reference used within a [`Node`] of the [Hypercore](hypercore::Hypercore) block where a
 /// key-value  pair is stored.
-pub struct KeyValue {
+struct KeyValue {
     /// Index of key value pair within the [`hypercore::Hypercore`].
     seq: u64,
 }
 
+#[derive(Clone, Debug)]
+/// Data related to a key value pair within the [`Hyperbee`].
 pub struct KeyValueData {
+    /// The index of the block within the [`Hypercore`](hypercore::Hypercore) where this data is stored.
     pub seq: u64,
+    /// The key. The data by which the [`Hyperbee`] is ordered.
     pub key: Vec<u8>,
+    /// The value.
     pub value: Option<Vec<u8>>,
 }
 
-// TODO make not pub
 #[derive(Debug)]
 /// Pointer used within a [`Node`] to reference to it's child nodes.
-pub struct Child<M: CoreMem> {
-    /// Index of the [`BlockEntry`]within the [`hypercore::Hypercore`] that contains the [`Node`]
+struct Child<M: CoreMem> {
+    /// Index of the [`BlockEntry`] within the [`hypercore::Hypercore`] that contains the [`Node`]
     pub seq: u64,
     /// Index of the `Node` within the [`BlockEntry`] referenced by [`Child::seq`]
     pub offset: u64,
@@ -76,11 +80,10 @@ pub struct Child<M: CoreMem> {
     cached_node: Option<SharedNode<M>>,
 }
 
-//TODO make not pub
 #[derive(Clone, Debug)]
 /// A "block" from a [`Hypercore`](hypercore::Hypercore) deserialized into the form used in
 /// Hyperbee
-pub struct BlockEntry<M: CoreMem> {
+struct BlockEntry<M: CoreMem> {
     /// Pointers::new(NodeSchema::new(hypercore.get(seq)).index))
     nodes: Vec<SharedNode<M>>,
     /// NodeSchema::new(hypercore.get(seq)).key
@@ -99,152 +102,12 @@ struct Children<M: CoreMem> {
     children: RwLock<Vec<Child<M>>>,
 }
 
-// TODO make not pub
 /// A node of the B-Tree within the [`Hyperbee`]
 #[derive(Debug)]
-pub struct Node<M: CoreMem> {
-    pub keys: Vec<KeyValue>,
+struct Node<M: CoreMem> {
+    keys: Vec<KeyValue>,
     children: Children<M>,
     blocks: Shared<Blocks<M>>,
-}
-
-/// An append only B-Tree built on [`Hypercore`](hypercore::Hypercore). It provides a key-value
-/// store API, with methods for [inserting](Hyperbee::put), [getting](Hyperbee::get), and
-/// [deleting](Hyperbee::del) key-value pair. As well as creating [sorted
-/// iterators](Hyperbee::traverse), and ["sub" B-Trees](Hyperbee::sub) for grouping related data.
-#[derive(Debug, Builder)]
-#[builder(pattern = "owned", derive(Debug))]
-pub struct Hyperbee<M: CoreMem> {
-    tree: Shared<Tree<M>>,
-}
-
-impl<M: CoreMem> Hyperbee<M> {
-    /// The number of blocks in the hypercore.
-    /// The first block is always the header block so:
-    /// `version` would be the `seq` of the next block
-    /// `version - 1` is most recent block
-    pub async fn version(&self) -> u64 {
-        self.tree.read().await.version().await
-    }
-    /// Gets the root of the tree.
-    /// When `ensure_header == true` write the hyperbee header onto the hypercore if it does not exist.
-    pub async fn get_root(
-        &self,
-        ensure_header: bool,
-    ) -> Result<Option<Shared<Node<M>>>, HyperbeeError> {
-        self.tree.read().await.get_root(ensure_header).await
-    }
-
-    /// Create the header for the Hyperbee. This must be done before writing anything else to the
-    /// tree.
-    pub async fn create_header(
-        &self,
-        metadata: Option<Metadata>,
-    ) -> Result<AppendOutcome, HyperbeeError> {
-        self.tree.read().await.create_header(metadata).await
-    }
-    /// Returs a string representing the structure of the tree showing the keys in each node
-    pub async fn print(&self) -> Result<String, HyperbeeError> {
-        self.tree.read().await.print().await
-    }
-
-    /// Get the value corresponding to the provided `key` from the Hyperbee
-    /// # Errors
-    /// When `Hyperbee.get_root` fails
-    pub async fn get(&self, key: &[u8]) -> Result<Option<(u64, Option<Vec<u8>>)>, HyperbeeError> {
-        self.tree.read().await.get(key).await
-    }
-
-    /// Insert the given key and value into the tree
-    /// Returs the `seq` of the new key, and `Option<u64>` which contains the `seq` of the old key
-    /// if it was replaced.
-    #[tracing::instrument(level = "trace", skip(self), ret)]
-    pub async fn put(
-        &self,
-        key: &[u8],
-        value: Option<&[u8]>,
-    ) -> Result<(Option<u64>, u64), HyperbeeError> {
-        self.tree.read().await.put(key, value).await
-    }
-
-    /// Like [`Hyperbee::put`] but takes a `compare_and_swap` function.
-    /// The `compared_and_swap` function is called with the old key (if present), and the new key.
-    /// The new key is only inserted if `compare_and_swap` returns true.
-    /// Returs two `Option<u64>`s. The first is the old key, the second is the new key.
-    pub async fn put_compare_and_swap(
-        &self,
-        key: &[u8],
-        value: Option<&[u8]>,
-        cas: impl FnOnce(Option<&KeyValueData>, &KeyValueData) -> bool,
-    ) -> Result<(Option<u64>, Option<u64>), HyperbeeError> {
-        self.tree
-            .read()
-            .await
-            .put_compare_and_swap(key, value, cas)
-            .await
-    }
-
-    /// Delete the given key from the tree.
-    /// Returns the `seq` from the key if it was deleted.
-    pub async fn del(&self, key: &[u8]) -> Result<Option<u64>, HyperbeeError> {
-        self.tree.read().await.del(key).await
-    }
-
-    /// Like [`Hyperbee::del`] but takes a `compare_and_swap` function.
-    /// Before deleting the function is called with existing key's [`KeyValueData`].
-    /// The key is only deleted if `compare_and_swap` returs true.
-    /// Returns the `bool` representing the result of `compare_and_swap`, and the `seq` for the
-    /// key.
-    pub async fn del_compare_and_swap(
-        &self,
-        key: &[u8],
-        cas: impl FnOnce(&KeyValueData) -> bool,
-    ) -> Result<Option<(bool, u64)>, HyperbeeError> {
-        self.tree.read().await.del_compare_and_swap(key, cas).await
-    }
-
-    /// Create a new tree with all it's operation's prefixed by the provided `prefix`.
-    pub fn sub(&self, prefix: &[u8], config: PrefixedConfig) -> Prefixed<M> {
-        Prefixed::new(prefix, self.tree.clone(), config)
-    }
-
-    /// Traverse the tree based on the given [`TraverseConfig`]
-    pub async fn traverse<'a>(
-        &self,
-        conf: TraverseConfig,
-    ) -> Result<Traverse<'a, M>, HyperbeeError> {
-        self.tree.read().await.traverse(conf).await
-    }
-}
-
-impl Hyperbee<random_access_disk::RandomAccessDisk> {
-    /// Helper for creating a Hyperbee
-    /// # Panics
-    /// when storage path is incorrect
-    /// when Hypercore failse to build
-    /// when Blocks fails to build
-    ///
-    /// # Errors
-    /// when Hyperbee fails to build
-    pub async fn from_storage_dir<T: AsRef<Path>>(
-        path_to_storage_dir: T,
-    ) -> Result<Hyperbee<random_access_disk::RandomAccessDisk>, HyperbeeError> {
-        let tree = tree::Tree::from_storage_dir(path_to_storage_dir).await?;
-        Ok(HyperbeeBuilder::default()
-            .tree(Arc::new(RwLock::new(tree)))
-            .build()?)
-    }
-}
-
-impl Hyperbee<random_access_memory::RandomAccessMemory> {
-    /// Helper for creating a Hyperbee in RAM
-    pub async fn from_ram(
-    ) -> Result<Hyperbee<random_access_memory::RandomAccessMemory>, HyperbeeError> {
-        let tree = tree::Tree::from_ram().await?;
-        Ok(HyperbeeBuilder::default()
-            .tree(Arc::new(RwLock::new(tree)))
-            .build()?)
-    }
 }
 
 impl KeyValue {
@@ -377,7 +240,6 @@ impl<M: CoreMem> Children<M> {
 ///         the index within this `node`'s keys where the `key` wolud be inserted
 /// if `matched` is Some:
 ///     the index within this `node`'s keys of the `key`
-// TODO rename me because it is not just child index
 async fn get_index_of_key<M: CoreMem, T>(
     node: SharedNode<M>,
     key: &T,